Move DOI field into own file
